requirements.txt
dj-database-url==0.4.1
Django==1.9.5
django-rest-auth==0.7.0
djangorestframework==3.3.3
httpie==0.9.3
psycopg2==2.6.1
Pygments==2.1.3
requests==2.10.0
six==1.10.0
